* Moving `ObservationStatus` to settings file.
